LUN-399: add grantPermissions hook, cleanUp ui
