+++ src/components/AudioRecordOverlay.tsx
@@ -1,10 +1,11 @@
-import React, { ReactElement, useCallback, useEffect, useMemo, useState } from 'react'
+import React, { ReactElement, useMemo, useState } from 'react'
 import { Modal as RNModal, Platform, Pressable } from 'react-native'
-import { PERMISSIONS, request, requestMultiple, RESULTS } from 'react-native-permissions'
+import { PERMISSIONS } from 'react-native-permissions'
 import { heightPercentageToDP as hp } from 'react-native-responsive-screen'
 import styled, { useTheme } from 'styled-components/native'
 
-import { CloseIcon, FloppyDiskIcon, MicrophoneIcon, PlayIcon, StopIcon } from '../../assets/images'
+import { CloseIcon, MicrophoneIcon } from '../../assets/images'
+import useGrantPermissions from '../hooks/useGrantPermissions'
 import { getLabels } from '../services/helpers'
 import { reportError } from '../services/sentry'
 import { HeadingText } from './text/Heading'
@@ -17,7 +18,6 @@ const Container = styled.SafeAreaView`
 const Icon = styled.Pressable`
   align-self: flex-end;
   margin: ${props => `${props.theme.spacings.xs} ${props.theme.spacings.md}`};
-  align-self: center;
 `
 
 const RecordIcon = styled(Pressable)<{ isPressed: boolean }>`
@@ -72,22 +72,12 @@ const InfoContainer = styled.View`
   justify-content: flex-end;
 `
 
-const IconContainer = styled.View`
-  justify-content: flex-end;
-  flex-direction: row;
-`
-
 interface AudioRecordOverlayProps {
   onClose: () => void
   onStartRecording: () => Promise<void>
   onStopRecording: () => Promise<void>
-  onStartPlay: () => Promise<void>
-  onStopPlay: () => Promise<void>
-  onSaveRecording: () => void
   recordingTime: string
   meteringResults: number[]
-  isPlaying: boolean
-  isRecording: boolean
 }
 
 // Zero alignment of values with the minimum metering value
@@ -100,77 +90,30 @@ const AudioRecordOverlay = ({
   onClose,
   onStartRecording,
   onStopRecording,
-  onStartPlay,
-  onStopPlay,
-  onSaveRecording,
   recordingTime,
   meteringResults,
-  isPlaying,
-  isRecording,
 }: AudioRecordOverlayProps): ReactElement => {
   const [isPressed, setIsPressed] = useState<boolean>(false)
-  const [permissionRequested, setPermissionRequested] = useState<boolean>(false)
-  const [permissionGranted, setPermissionGranted] = useState<boolean>(false)
+  const permissionGranted = useGrantPermissions(
+    Platform.OS === 'android'
+      ? [
+          PERMISSIONS.ANDROID.RECORD_AUDIO,
+          PERMISSIONS.ANDROID.READ_EXTERNAL_STORAGE,
+          PERMISSIONS.ANDROID.WRITE_EXTERNAL_STORAGE,
+        ]
+      : PERMISSIONS.IOS.MICROPHONE
+  )
   const { hold, talk } = getLabels().general.audio
   const theme = useTheme()
-
-  const requestAndroidPermissions = useCallback(() => {
-    requestMultiple([
-      PERMISSIONS.ANDROID.RECORD_AUDIO,
-      PERMISSIONS.ANDROID.READ_EXTERNAL_STORAGE,
-      PERMISSIONS.ANDROID.WRITE_EXTERNAL_STORAGE,
-    ])
-      .then(result => {
-        const grants = Object.values(result)
-        setPermissionGranted(grants.filter((el, i) => grants.indexOf(el) === i).toString() === RESULTS.GRANTED)
-      })
-      .catch(reportError)
-      .finally(() => setPermissionRequested(true))
-  }, [])
-
-  const requestIosPermission = useCallback(() => {
-    request(PERMISSIONS.IOS.MICROPHONE)
-      .then(result => setPermissionGranted(result === RESULTS.GRANTED))
-      .catch(reportError)
-      .finally(() => setPermissionRequested(true))
-  }, [])
-
-  useEffect(() => {
-    if (!permissionRequested) {
-      if (Platform.OS === 'android') {
-        requestAndroidPermissions()
-      } else {
-        requestIosPermission()
-      }
-    }
-  }, [permissionRequested, requestIosPermission, requestAndroidPermissions])
 
   const cleanedMetering = useMemo(() => cleanUpMeteringResults(meteringResults), [meteringResults])
 
   return (
     <RNModal visible transparent animationType='fade' onRequestClose={() => onClose()}>
       <Container>
-        <IconContainer>
-          {!isRecording && meteringResults.length > 0 && (
-            <>
-              {isPlaying ? (
-                <Icon onPress={() => onStopPlay()}>
-                  <StopIcon width={hp('3.5%')} height={hp('3.5%')} />
-                </Icon>
-              ) : (
-                <Icon onPress={() => onStartPlay()}>
-                  <PlayIcon width={hp('3.5%')} height={hp('3.5%')} />
-                </Icon>
-              )}
-              <Icon onPress={() => onSaveRecording()}>
-                <FloppyDiskIcon width={hp('3.5%')} height={hp('3.5%')} />
-              </Icon>
-            </>
-          )}
-          <Icon onPress={() => onClose()}>
-            <CloseIcon width={theme.spacingsPlain.lg} height={theme.spacingsPlain.lg} />
-          </Icon>
-        </IconContainer>
+        <Icon onPress={() => onClose()}>
+          <CloseIcon width={theme.spacingsPlain.lg} height={theme.spacingsPlain.lg} />
+        </Icon>
         <Content>
           <HeadingContainer>
             <Heading>{isPressed ? talk : hold}</Heading>
